requirements-recognizer.txt
numpy
pillow
torch
opencv-python-headless
